requirements.txt
python-dotenv
streamlit
openai
langchain
pypdf
chromadb
PyPDF2
watchdog
faiss-cpu
